add swagger description & auth check

+++ src/main/java/spring/board/web/QuestionController.java
@@ -1,5 +1,6 @@
 package spring.board.web;
 
+import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -8,6 +9,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import spring.board.domain.Question;
+import spring.board.exception.ApiExceptions;
+import spring.board.exception.ErrorCode;
 import spring.board.repository.UserRepository;
 import spring.board.service.QuestionService;
 import spring.board.web.dto.QuestionForm;
@@ -26,7 +29,8 @@ public class QuestionController {
     }
 
     @GetMapping("/list/{userid}")
-    public ResponseEntity<Page<Question>> showByUser(@PathVariable String userid, Pageable pageable, Authentication auth) {
+    public ResponseEntity<Page<Question>> showByUser(@Parameter(name = "userid", description = "user's email") @PathVariable String userid,
+                                                     Pageable pageable, Authentication auth) {
         String email = auth.getName();
 
         if (!email.equals(userid)) return null;
@@ -36,7 +40,9 @@ public class QuestionController {
     }
 
     @GetMapping("/viewOne/{id}")
-    public ResponseEntity<Question> viewOne(@PathVariable Long id) {
+    public ResponseEntity<Question> viewOne(@Parameter(name = "id", description = "question's id") @PathVariable Long id,
+                                            Authentication auth) {
+        if (!questionService.checkUserAvailable(id, auth.getName())) throw new ApiExceptions(ErrorCode.MEMBER_NOT_ALLOWED);
         return ResponseEntity.ok(questionService.viewOne(id));
     }
 
@@ -46,12 +52,15 @@ public class QuestionController {
     }
 
     @PostMapping(value = "/edit/{id}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<Question> edit(@PathVariable Long id, @ModelAttribute QuestionForm form) {
+    public ResponseEntity<Question> edit(@Parameter(name = "id", description = "question's id") @PathVariable Long id, @ModelAttribute QuestionForm form,
+                                         Authentication auth) {
+        if (!questionService.checkUserAvailable(id, auth.getName())) throw new ApiExceptions(ErrorCode.MEMBER_NOT_ALLOWED);
         return ResponseEntity.ok(questionService.updateQuestion(id, form));
     }
 
     @DeleteMapping("/del/{id}")
-    public ResponseEntity<Boolean> deleteQuestion(@PathVariable Long id) {
+    public ResponseEntity<Boolean> deleteQuestion(@Parameter(name = "id", description = "question's id") @PathVariable Long id, Authentication auth) {
+        if (!questionService.checkUserAvailable(id, auth.getName())) throw new ApiExceptions(ErrorCode.MEMBER_NOT_ALLOWED);
         return ResponseEntity.ok(questionService.deleteQuestion(id));
     }
 }
